pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["Ilya Malahov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"
kombu = "5.3.6"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-beat = "^2.6.0"
gevent = "^24.2.1"
eventlet = "^0.36.1"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
django-redis = "^5.4.0"
rest-framework-simplejwt = "^0.0.2"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"